Add files via upload
I was given a csv file with a list of all the transactions that took place for the MLB season. I parsed through the transactions to create a dataframe that specified the injury type, body parts and side( right or left). I then exported my results into a csv file. Now users are able to filter through injury types and body parts